Move lifetime related functions out of phasor module
