+++ docs/api/lifetime.rst
@@ -1,0 +1,5 @@
+phasorpy.lifetime
+-----------------
+
+.. automodule:: phasorpy.lifetime
+    :members:
