+++ src/phasorpy/plot/_lifetime_plots.py
@@ -17,10 +17,12 @@ from matplotlib.lines import Line2D
 from matplotlib.widgets import Slider
 
 from .._utils import update_kwargs
-from ..phasor import (
+from ..lifetime import (
     lifetime_to_frequency,
     lifetime_to_signal,
     phasor_from_lifetime,
+)
+from ..phasor import (
     phasor_to_polar,
     phasor_transform,
 )
