+++ tutorials/api/phasorpy_filtering.py
@@ -15,8 +15,8 @@ include the median and wavelet filtering.
 
 from phasorpy.datasets import fetch
 from phasorpy.io import signal_from_imspector_tiff
+from phasorpy.lifetime import phasor_calibrate
 from phasorpy.phasor import (
-    phasor_calibrate,
     phasor_filter_median,
     phasor_filter_pawflim,
     phasor_from_signal,
